requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.13.1
altair==4.2.2
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
astroid==3.0.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.2
bitstring==4.0.2
bs4==0.0.1
cachetools==5.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
coloredlogs==14.0
comtypes==1.1.14
contourpy==1.0.7
croniter==1.3.8
cryptography==41.0.7
cycler==0.11.0
dagit==1.2.2
dagster==1.2.2
dagster-graphql==1.2.2
dill==0.3.8
distro==1.9.0
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
docstring-parser==0.15
ecdsa==0.18.0
email-validator==2.1.0.post1
entrypoints==0.4
esptool==4.6.2
et-xmlfile==1.1.0
fastapi==0.108.0
ffmpeg==1.4
ffmpy==0.3.0
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.3.0
gast==0.4.0
google-auth==2.17.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gql==3.4.0
gradio==3.24.1
gradio_client==0.0.8
graphene==3.2.2
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
grpcio==1.51.3
grpcio-health-checking==1.51.3
h11==0.14.0
h5py==3.8.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
linkify-it-py==2.0.0
Mako==1.3.1
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
ml-dtypes==0.0.4
multidict==6.0.4
neuralintents==0.0.4
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
openai==1.8.0
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.10
packaging==23.0
pandas==1.5.3
passlib==1.7.4
pendulum==2.1.2
Pillow==9.5.0
platformdirs==4.2.0
protobuf==4.22.1
psutil==5.9.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.2.8
PyAudio==0.2.13
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
pydub==0.25.1
PyJWT==2.8.0
pylint==3.0.3
pyparsing==3.0.9
pypiwin32==223
pyreadline3==3.4.1
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pyttsx3==2.90
pytz==2022.7.1
pytzdata==2020.1
pywin32==305
PyYAML==6.0.1
reedsolo==1.7.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rsa==4.9
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4
SpeechRecognition==3.10.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
starlette==0.32.0.post1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
toposort==1.10
tqdm==4.65.0
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.1
ujson==5.9.0
universal_pathlib==0.0.22
urllib3==1.26.15
uvicorn==0.25.0
watchdog==2.3.1
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
Werkzeug==2.2.3
wrapt==1.14.1
wsl==0.12
xlrd==2.0.1
yarl==1.8.2
